[model][client] Model registration
To support type custom fields and richer metadata modeling, we implement
the registerModels feature of the SDK to convert @MClass models into JSON
to be registered with the server. Fields are limited to numeric types,
Boolean, Instant, and String. We allow for a TEXT field. We convert enum
types into a String field (enum support to be added later).